--- a/lib/client.js
+++ b/lib/client.js
@@ -140,6 +140,16 @@ class Client extends events.EventEmitter {
     }
 
     /**
+     * Manually prepare query into prepared statement
+     * @param {string} query
+     * @returns {Promise<rust.PreparedStatementWrapper>}
+     * @package
+     */
+    async prepareQuery(query) {
+        return await this.rustClient.prepareStatement(query);
+    }
+
+    /**
      * Attempts to connect to one of the [contactPoints]{@link ClientOptions} and discovers the rest the nodes of the
      * cluster.
      *
@@ -256,7 +266,7 @@ class Client extends events.EventEmitter {
         try {
             const execOptions = this.createOptions(options);
             return promiseUtils.optionalCallback(
-                this.#rustyExecute(query, params, execOptions),
+                this.rustyExecute(query, params, execOptions),
                 callback,
             );
         } catch (err) {
@@ -658,13 +668,13 @@ class Client extends events.EventEmitter {
 
     /**
      * Wrapper for executing queries by rust driver
-     * @param {string} query
+     * @param {string | rust.PreparedStatementWrapper} query
      * @param {Array} params
      * @param {ExecOptions.ExecutionOptions} execOptions
      * @returns {Promise<ResultSet>}
-     * @private
+     * @package
      */
-    async #rustyExecute(query, params, execOptions) {
+    async rustyExecute(query, params, execOptions) {
         if (
             // !execOptions.isPrepared() &&
             params &&
@@ -709,7 +719,7 @@ class Client extends events.EventEmitter {
      * Core part of executing rust queries
      * @param {ResolveCallback} resolve
      * @param {RejectCallback} reject
-     * @param {string} query
+     * @param {string | rust.PreparedStatementWrapper} query
      * @param {Array} params
      * @param {ExecOptions.ExecutionOptions} execOptions
      */
@@ -719,7 +729,10 @@ class Client extends events.EventEmitter {
             let result;
             if (execOptions.isPrepared()) {
                 // Execute prepared statement, as requested by the user
-                let statement = await this.rustClient.prepareStatement(query);
+                let statement =
+                    query instanceof rust.PreparedStatementWrapper
+                        ? query
+                        : await this.rustClient.prepareStatement(query);
                 let parsedParams = parseParams(
                     statement.getExpectedTypes(),
                     params,
@@ -730,6 +743,11 @@ class Client extends events.EventEmitter {
                     rustOptions,
                 );
             } else {
+                if (query instanceof rust.PreparedStatementWrapper) {
+                    throw new Error(
+                        "Unexpected prepared statement wrapper for unprepared queries",
+                    );
+                }
                 let expectedTypes = convertHints(execOptions.getHints() || []);
                 let parsedParams = parseParams(expectedTypes, params, true);
                 result = await this.rustClient.queryUnpaged(