--- a/lib/client.js
+++ b/lib/client.js
@@ -18,6 +18,7 @@ const rust = require("../index");
 const ResultSet = require("./types/result-set.js");
 const { parseParams, convertHints } = require("./types/cql-utils.js");
 const queryOptions = require("./query-options.js");
+const { PreparedCache } = require("./cache.js");
 
 /**
  * Represents a database client that maintains multiple connections to the cluster nodes, providing methods to
@@ -567,6 +568,7 @@ class Client extends events.EventEmitter {
         let allQueries = [];
         let parametersRows = [];
         let hints = execOptions.getHints() || [];
+        let preparedCache = new PreparedCache();
 
         for (let i = 0; i < queries.length; i++) {
             let element = queries[i];
@@ -576,17 +578,24 @@ class Client extends events.EventEmitter {
             /**
              * @type {rust.PreparedStatementWrapper | string}
              */
-            let query = typeof element === "string" ? element : element.query;
+            let statement =
+                typeof element === "string" ? element : element.query;
             let params = element.params || [];
             let types;
 
-            if (!query) {
+            if (!statement) {
                 throw new errors.ArgumentError(`Invalid query at index ${i}`);
             }
 
             if (shouldBePrepared) {
-                query = await this.rustClient.prepareStatement(query);
-                types = query.getExpectedTypes();
+                let prepared = preparedCache.getElement(statement);
+                if (!prepared) {
+                    prepared =
+                        await this.rustClient.prepareStatement(statement);
+                    preparedCache.storeElement(statement, prepared);
+                }
+                types = prepared.getExpectedTypes();
+                statement = prepared;
             } else {
                 types = convertHints(hints[i] || []);
             }
@@ -594,7 +603,7 @@ class Client extends events.EventEmitter {
             if (params) {
                 params = parseParams(types, params, shouldBePrepared === false);
             }
-            allQueries.push(query);
+            allQueries.push(statement);
             parametersRows.push(params);
         }