--- a/lib/concurrent/index.js
+++ b/lib/concurrent/index.js
@@ -5,6 +5,7 @@ const utils = require("../utils");
 const { Stream } = require("stream");
 const { Mutex } = require("async-mutex");
 const { env } = require("process");
+const { PreparedCache } = require("../cache");
 
 /**
  * Utilities for concurrent query execution with the DataStax Node.js Driver.
@@ -123,9 +124,10 @@ class ArrayBasedExecutor {
         });
         this._result = new ResultSetGroup(options);
         this._stop = false;
+        this._cache = new PreparedCache();
     }
 
-    execute() {
+    async execute() {
         const promises = new Array(this._concurrencyLevel);
 
         for (let i = 0; i < this._concurrencyLevel; i++) {
@@ -135,7 +137,7 @@ class ArrayBasedExecutor {
         return Promise.all(promises).then(() => this._result);
     }
 
-    _executeOneAtATime(initialIndex, iteration) {
+    async _executeOneAtATime(initialIndex, iteration) {
         const index = initialIndex + this._concurrencyLevel * iteration;
 
         if (index >= this._parameters.length || this._stop) {
@@ -154,8 +156,14 @@ class ArrayBasedExecutor {
             params = item;
         }
 
+        let prepared = this._cache.getElement(query);
+        if (!prepared) {
+            prepared = await (this._client.prepareQuery(query));
+            this._cache.storeElement(query, prepared);
+        }
+
         return this._client
-            .execute(query, params, this._queryOptions)
+            .rustyExecute(prepared, params, this._queryOptions)
             .then((rs) => this._result.setResultItem(index, rs))
             .catch((err) => this._setError(index, err))
             .then(() => this._executeOneAtATime(initialIndex, iteration + 1));