Cache prepared statements on the JS side within single call of batch and concurrent execute.

About 20% improvement on concurrent insert (both task clock and time elapsed).

Before (js-caching branch):

```bash
sudo perf stat node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000

 Performance counter stats for 'node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000':

         69,564.54 msec task-clock                       #    1.639 CPUs utilized             
         2,702,067      context-switches                 #   38.843 K/sec                     
            24,744      cpu-migrations                   #  355.698 /sec                      
           190,975      page-faults                      #    2.745 K/sec                     
   283,675,180,415      cycles                           #    4.078 GHz                       
    75,943,317,270      stalled-cycles-frontend          #   26.77% frontend cycles idle      
   238,391,112,023      instructions                     #    0.84  insn per cycle            
                                                  #    0.32  stalled cycles per insn   
    45,454,104,968      branches                         #  653.409 M/sec                     
     1,559,027,745      branch-misses                    #    3.43% of all branches           

      42.448039778 seconds time elapsed

      47.783400000 seconds user
      23.049097000 seconds sys
```

After:

```bash
sudo perf stat node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000

 Performance counter stats for 'node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000':

         54,311.66 msec task-clock                       #    1.619 CPUs utilized             
         1,910,653      context-switches                 #   35.179 K/sec                     
            18,347      cpu-migrations                   #  337.810 /sec                      
           213,348      page-faults                      #    3.928 K/sec                     
   225,277,325,784      cycles                           #    4.148 GHz                       
    59,589,673,035      stalled-cycles-frontend          #   26.45% frontend cycles idle      
   203,600,992,694      instructions                     #    0.90  insn per cycle            
                                                  #    0.29  stalled cycles per insn   
    38,835,678,700      branches                         #  715.052 M/sec                     
     1,230,468,744      branch-misses                    #    3.17% of all branches           

      33.556679150 seconds time elapsed

      37.259389000 seconds user
      17.954469000 seconds sys

```